// tests/types/organization_add.ts
export interface OrganizationData_add {
    companyName: string;
    locationName: string;
    shortName: string;
    address1: string;
    address2: string;
    city: string;
    stateProvince: string;
    postalCode: string;
    country: string;
    industry: string;
    linkedIn: string;
    website: string;
    parentOrg: string;
    orgType: string;
    status: string;
    phoneNumber: string;
    email: string;
    consultant: string;
    orgLogo: string;
    invoiceFrequency: string;
    belongWM: boolean;
    redWM: boolean;
}